fix: apply Stage 0 hardening review fix wave
Ship-after-fixes verdict from the whole-branch review: docs drift, one missing
test, two cosmetics — no correctness/security defects.

- dev/run-server.sh: chmod 600 dev/server.yaml before starting, so the new
  permission warning doesn't cry wolf on every fresh clone (the committed file
  is mode 0644; git can't carry 0600).
- client/zig/CLAUDE.md: document --version (handled pre-config-load),
  non-409 error surfacing via parseServerError/defaultMessage, and the
  git-checkout build requirement (-Dversion= override).
- CLAUDE.md, dev/README.md: document the bind and rate_limit config keys.
- server/main.go: extract validateConfig (empty-secret + rate-limit
  resolution) out of main so it's unit-testable; move warnIfPermissive before
  the fatal check so both problems surface in one pass. server/main_test.go:
  cover it, pinning the existing whitespace-secret-accepted behavior.
- server/store_test.go: rename TestAbsentDataFormatVersionAssumedCurrent to
  TestAbsentDataFormatVersionTreatedAsV1 (assertion was already correct, name
  was misleading).
- server/task.go: State.DataFormatVersion uint -> uint64, matching
  StateVersion.
- client/zig/src/api/client.zig: correct fail()'s comment, which claimed its
  narrow error{Reported} return type protects run()'s error set — no longer
  true since stdoutErr widens run() via anyerror.

+++ server/task.go
@@ -73,7 +73,7 @@ const CurrentDataFormatVersion = 1
 
 // State is the whole server state and the on-disk JSON shape.
 type State struct {
-	DataFormatVersion uint            `json:"data_format_version"`
+	DataFormatVersion uint64          `json:"data_format_version"`
 	StateVersion      uint64          `json:"state_version"`
 	Tasks             map[string]Task `json:"tasks"`
 }
